Реализация рэйкастинга для AABB в произвольной позиции

+++ src/voxels/Chunks.h
@@ -48,8 +48,8 @@ public:
 				   glm::vec3 dir, 
 				   float maxLength, 
 				   glm::vec3& end, 
-				   glm::vec3& norm, 
-				   glm::vec3& iend);
+				   glm::ivec3& norm, 
+				   glm::ivec3& iend);
 
 	const AABB* isObstacle(float x, float y, float z);
